config/settings.py
import os
from dotenv import load_dotenv

load_dotenv()  # load .env file at startup

GROQ_API_KEY = os.getenv("GROQ_API_KEY")
TAVILY_API_KEY = os.getenv("TAVILY_API_KEY")
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")  # optional
